build: add version, files, and license to package.json
This commit introduces several important updates to the `package.json`. A specific version (0.1.0) has been specified, which marks the initial step toward proper version management and package distribution. Furthermore, the inclusion of the "files" field, specifying the "./lib" directory, ensures that only the necessary files are packaged, reducing the package size and improving load times. Additionally, defining the "license" field as "MIT" clearly communicates the licensing terms under which this package can be used, which is crucial for fostering a transparent and open-source ecosystem. Overall, these changes aim to enhance the package's usability and compliance with best practices in package management.