chore: Bump version to 0.1.2 in package.json
Updated the version in package.json from 0.1.1 to 0.1.2. This increment reflects changes or improvements not specific to any particular aspect of the project but likely minor enhancements or minor bug fixes. This version change is part of ongoing maintenance and should not affect project compatibility or require modifications for project consumers.